requirements.txt
certifi==2019.3.9
chardet==3.0.4
fake-useragent==0.1.11
gunicorn==19.9.0
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
urllib3==1.24.1
selenium==3.141.0
pillow==6.0.0
oauth2client==4.1.3
